Update Styling to generate in `theme` dir. Move PathService to its own folder and separate DomainPathService into its own file.
