+++ lib/generation/flutter_project_builder/post_gen_tasks/global_styling/text_styles_post_gen_task.dart
@@ -1,7 +1,9 @@
+import 'package:get_it/get_it.dart';
 import 'package:parabeac_core/controllers/main_info.dart';
 import 'package:parabeac_core/generation/flutter_project_builder/post_gen_tasks/post_gen_task.dart';
 import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy/commands/add_constant_command.dart';
 import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy/file_ownership_policy.dart';
+import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy/path_services/path_service.dart';
 import 'package:parabeac_core/generation/generators/value_objects/generation_configuration/pb_generation_configuration.dart';
 import 'package:parabeac_core/generation/generators/visual-widgets/pb_text_gen.dart';
 import 'package:pbdl/pbdl.dart';
@@ -37,6 +39,7 @@ class TextStylesPostGenTask extends PostGenTask {
         filename: '${mainInfo.projectName.snakeCase}_text_styles',
         ownershipPolicy: FileOwnership.PBC,
         imports: 'import \'package:flutter/material.dart\';',
+        relativePath: GetIt.I.get<PathService>().themingRelativePath,
       ),
     );
   }
