+++ lib/generation/generators/value_objects/file_structure_strategy/commands/responsive_layout_builder_command.dart
@@ -1,7 +1,7 @@
 import 'dart:collection';
 import 'package:get_it/get_it.dart';
 import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy/file_ownership_policy.dart';
-import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy/path_service.dart';
+import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy/path_services/path_service.dart';
 import 'package:path/path.dart' as p;
 import 'package:parabeac_core/controllers/main_info.dart';
 import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy/commands/file_structure_command.dart';
